Clicking 'edit' button opens todo dialog In order not to repeat myself, the 'edit todo' dialog is the 'add todo' dialog with a different heading and submit button. The event listener for each button shows or hides the corresponding elements when clicked.
The idea is that clicking 'edit' takes that todo's ID and fills in all
the fields in the todo dialog. User can then make changes and submit.
The submit button then sends the changes to the todo. (so it changes
the existing todo, it doesn't make a new one).

How to?:
- populate fields based on existing todo id?
Do:
- get todo's ID from click: e.target.id
- create a new method on todo class that changes all properties at once